adding new processing of segments

+++ servers/common/service/data/models/result.go
@@ -1,7 +1,7 @@
 package models
 
 type Result struct {
-	SegmentId   int32   `json:"segment_id" bson:"segment_id"`
+	SegmentId   int64   `json:"segment_id" bson:"segment_id"`
 	Time        float64 `json:"time" bson:"time"`
 	Rank        int     `json:"rank" bson:"rank"`
 	ActivityId  int64   `json:"activity_id" bson:"activity_id"`
