+++ servers/common/service/data/models/stage.go
@@ -3,7 +3,7 @@ package models
 import "github.com/cjburchell/strava-go"
 
 type Stage struct {
-	SegmentId    int32               `json:"segment_id" bson:"segment_id"`
+	SegmentId    int64               `json:"segment_id" bson:"segment_id"`
 	Distance     float64             `json:"distance" bson:"distance"`
 	Number       int                 `json:"number" bson:"number"`
 	ActivityType strava.ActivityType `json:"activity_type" bson:"activity_type"`
